requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
streamlit
fastapi
uvicorn
missingno 
pyarrow
langgraph
langchain 
langchain-core
pydantic
llamaapi
python-dotenv
langchain-community